Show ping for autopunch hosts
This simple patch simply delays the PingMan initialisation to when the
network menu appears.

This makes PingMan run its "bind" syscall for the ping socket a few
seconds after the game starts (while the game loads and the user opens
the VS network menu), which makes it run after autopunch has
initialized. This way, when we bind the socket, autopunch can properly
recognize it and inject its magic in ping messages.

Previously, we ran the bind call on game start, before autopunch had
time to check for updates and initialize itself. This meant that the
first bind call wasn't seen by autopunch, and autopunch then ignored the
next calls made on this socket, therefore not injecting hole-punching in
ping messages.

So delaying the PingMan init enables autopunch to properly detect the
socket, and show the ping for autopunch hosts. Tested locally.